+++ src/main/kotlin/org/osrsxp/service/Skill.kt
@@ -1,3 +1,3 @@
 package org.osrsxp.service
 
-class Skill(val name: String, val rank: Long, val level: Long, val xp: Long) {}
+class Skill(val name: String, val rank: Long, val level: Int, val xp: Long) {}
